+++ react/Address.tsx
@@ -1,14 +1,21 @@
 import React, { FunctionComponent } from 'react'
 import { AddressRules, AddressSummary } from 'vtex.address-form'
+import { useCssHandles } from 'vtex.css-handles'
 
 interface Props {
   address: Address
   pickup?: Parcel
 }
 
+const CSS_HANDLES = [
+  'addressContainer'
+] as const
+
 const Address: FunctionComponent<Props> = ({ address, pickup }) => {
+
+  const handles = useCssHandles(CSS_HANDLES)
   return (
-    <div className="c-muted-1 lh-copy" data-testid="address-component">
+    <div className={`${handles.addressContainer} c-muted-1 lh-copy`} data-testid="address-component">
       {pickup && (
         <p className="c-on-base lh-copy">{pickup.pickupFriendlyName}</p>
       )}